requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
grpcio==1.70.0
grpcio-tools==1.70.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
protobuf==5.29.3
PyJWT==2.10.1
redis==5.2.1
setuptools==75.8.2
Werkzeug==3.1.3
